requirements/common.txt
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
psycopg2==2.7.1